requirements.txt
# ML и компьютерное зрение
torch>=1.9.0
torchvision>=0.10.0
yolov5>=6.1.0
easyocr>=1.4.1
opencv-python>=4.5.3.56
numpy>=1.21.0
Pillow>=8.3.1

# GUI
PySide6>=6.2.0

# API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.7.0

# Очереди и асинхронность
pika>=1.2.0
aio-pika>=7.1.0
asyncio>=3.4.3

# Обработка данных
pandas>=1.3.0
openpyxl>=3.0.7
python-dateutil>=2.8.2

# Логирование и мониторинг
prometheus-client>=0.11.0
python-json-logger>=2.0.2

# Тестирование
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=2.12.1
black>=21.7b0
isort>=5.9.3
mypy>=0.910

# Утилиты
python-dotenv>=0.19.0
pyyaml>=5.4.1
tqdm>=4.62.0
requests>=2.26.0

# Google Cloud Vision (опционально)
google-cloud-vision>=2.4.2 